Refactored project structure, added annotation mapper class
