+++ src/main/java/liquibase/ext/tests/utils/LiquibaseTask.java
@@ -1,4 +1,4 @@
-package liquibase.ext.tests;
+package liquibase.ext.tests.utils;
 
 public enum LiquibaseTask {
