+++ src/main/java/liquibase/ext/tests/utils/LiquibaseTaskLauncher.java
@@ -1,17 +1,12 @@
-package liquibase.ext.tests;
+package liquibase.ext.tests.utils;
 
 import liquibase.ext.tests.annotations.LiquibaseTest;
 import liquibase.ext.tests.exceptions.LiquibaseMigrationRunException;
 import liquibase.integration.commandline.Main;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 
 public class LiquibaseTaskLauncher {
-
-    private static final String ARG_PREFIX = "--";
-    private static final String ARG_EQUAL = "=";
-    private static final String ARG_CMD_CHANGE_LOG_FILE = ARG_PREFIX + "changeLogFile";
 
     private static final String DEFAULT_MESSAGE = "Can't run Liquibase migrations during tests";
 
@@ -34,15 +29,8 @@ public class LiquibaseTaskLauncher {
     }
 
     public static void update(LiquibaseTest annotation) {
-        String[] args = parseAnnotationSettings(annotation);
+        String[] args = AnnotationMapper.parseAnnotationSettings(annotation);
         update(args);
     }
 
-    private static String[] parseAnnotationSettings(LiquibaseTest annotation) {
-        ArrayList<String> settings = new ArrayList<>();
-        if (annotation.changeLogFile().length() > 0) {
-            settings.add(ARG_CMD_CHANGE_LOG_FILE + ARG_EQUAL + annotation.changeLogFile());
-        }
-        return settings.toArray(new String[settings.size()]);
-    }
 }
