+++ src/test/java/liquibase/ext/tests/LiquibaseTaskLauncherTest.java
@@ -1,9 +1,9 @@
 package liquibase.ext.tests;
 
-import liquibase.exception.CommandLineParsingException;
 import liquibase.exception.LiquibaseException;
 import liquibase.ext.tests.annotations.LiquibaseTest;
 import liquibase.ext.tests.exceptions.LiquibaseMigrationRunException;
+import liquibase.ext.tests.utils.LiquibaseTaskLauncher;
 import liquibase.integration.commandline.Main;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -11,7 +11,6 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import java.io.IOException;
 import java.lang.annotation.Annotation;
 
 import static org.mockito.Matchers.any;
