+++ src/test/java/liquibase/ext/tests/listeners/LiquibaseJUnitTestListenerTest.java
@@ -1,15 +1,11 @@
 package liquibase.ext.tests.listeners;
 
-import liquibase.ext.tests.LiquibaseTaskLauncher;
+import liquibase.ext.tests.utils.LiquibaseTaskLauncher;
 import liquibase.ext.tests.annotations.LiquibaseTest;
-import liquibase.ext.tests.examples.TestClassWithoutMigrations;
 import org.junit.Test;
 import org.junit.runner.Description;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PowerMockIgnore;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
